* Importar imágenes desde archivos NCER * Arreglado problema con archivos No_Tiled * Soporte al juego LAYTON2 (archivos .arc y plz) * Solucionado problema con archivos NCER

+++ Plugins/Images/Images/ntft.cs
@@ -63,7 +63,7 @@ namespace Images
 			ncgr.rahc.tileData = new NTFT();
 			ncgr.rahc.tileData.nPaleta = new byte[ncgr.rahc.nTiles];
 			ncgr.rahc.tileData.tiles = new byte[1][];
-            ncgr.rahc.tileData.tiles[0] = br.ReadBytes(ncgr.rahc.nTiles);
+            ncgr.rahc.tileData.tiles[0] = br.ReadBytes((int)ncgr.rahc.nTiles);
 
 			for (int i = 0; i < ncgr.rahc.nTiles; i++)
 			{
